Abort should abort a request regardless of whether we have a PVTypeInfo or not.
Bruce ran into an issue where we have a pending request for a
non-existent PV for which we have a valid typeinfo. In normal
circumstances, we would have completed the archive workflow, seen the
typeinfo and aborted the request. But since the PV does not exist, this
will never happen and we'll be stuck with this request in the queue.