Use a slightly more efficient method for filter integration.
The previous method based on unarchivedPV's was taking too much time as
it was checking the pvTypeInfo in detail. Instead, this new method
assumes that the number of names coming in is reasonably small; so we
pass it as is to all the appliances in the cluster and the lookup the
name in the cached copy of PVsForThisAppliance.