Upgrade hz to 3.7.1
This is an attempt to fix various "Not able to setup owner connection"
issues. Note that these errors may have nothing to do with this upgrade.
They may be related to KeepAlive settings on a proxied appliance.
Currently testing -Dhttp.keepAlive=false but also want to have this
upgrade ready in case this does not help.