Fix infinite loop issue when archiving large V4 PV's.
The default serialization control had hardcoded buffer sizes. It took
some juggling around to cater to larger arrays. ByteBuffer's can't be
subclasses that easily; so I had to use a ByteArrayOutputStream as a
temporary store. 

All this probably results in about 4-5 copies of the same data so this
is woefully inefficient. Next time we meet Matej/Michael, we should ask
for advice.