Get rid of the scan threads.
Gaudenz is consistently encountering a situation where SCAN Pv's are
dropping samples even though the buffer sizes are more than adequate. My
suspicion is that the SCAN threads are not arriving in time; and to rely
on this is probably not reliable anyways. So, the scan threads have been
eliminated; we now write the sample into the buffer if the timestamps
are newer that the last recorded value by more than the scan period.