Support for circular references when proxying data from other appliances
This should support cluster A proxying cluster B and cluster B proxying
cluster A. However, note that the appliance identities in the two
clusters have to be different. We use the appliance identity to break
any circular proxying effects.

Reworked this from the last commit to use skipExternalServers
similar to what the data retrieval uses to break circular references.