Proper support for archiveFields in policy files
This is https://github.com/slacmshankar/epicsarchiverap/issues/69.
Please see issue for more details.

This change fixes the archive PV workflow such that
1) Only those archiveFields that are in
getFieldsArchivedAsPartOfStream are added to the extraFields in the .VAL's PVTypeInfo
2) The other fields are converted to archive requests so that they
become their own PVTypeInfos.

There's a new unit test to test this - ArchiveFieldsNotInStreamTest.

There's no BPL to fix existing PVTypeInfo's as of yet.
Per Eric, he is ok to deleting these PV's and then readding the PV's
back to exercise the policy again. That seems to be a decent solution.