Use a bitset to check for timestamp changes.
PVAccess will not set the bit for the containing structure if an
sub-element of the structure changes. This should really be bit 0 and
all parent bits. However, I make the assumption that the timestamp field
is a field in the top level structure. So, we add bit 0 and then the
nits for the timestamp structure and its children. We use the interescts
call to check if the timestamp has changed.