Fire an event on the PV when a value is written into the stores.
For V3 PVs, this does not do anything.
For V4 PVs, this clears any changed fields. Our policy for storing
metafields into the "next" proper event needs some extra help for V4
where we can get full updates even on a field change. So we gather
metafield changes till a proper value is written into the store. This
should work for PV's that change frequently; but for changes that change
slowly the metafields will be written on the next record processing.