Support for append+alias.
Say we have a PV that changes it's name.
Because of the way the IOC deployment sometimes works; the new name is
added to the archiver before we have a chance to rename the old PV to
the new name. So now we have data with the old PV name, which no longer
exists and some data with the new name. Simply adding an alias means we
lose the old data.

So, this BPL
1) Appends the data for the old PV into the new PV
2) Deletes the old PV
3) Adds an alias for the old PV pointing to the new PV.

Note this is a fairly complex BPL; so ideally, one should backup the
data and do this in small batches. It is possible to lose data with this
BPL; so please do backup the database and the data.