Start time and end time as part of conversion
This is needed for the failover merge operation.
This is mostly a hint to the conversion functions as to the boundaries of the stream.
The DBR type conversion functions do not use this and convert every sample in the stream.
The failover merge conversion functions do us this as the from and to parameters for the retrieval request.