Add retrieval for retired PV without template
New algorithm allowing data for a retired PV to be retrieved
without the use of RetiredPVTemplate parameter.

If the PVTypeInfo is null because data for a retired PV was
requested, then perform a further search to see if data for
this PV exists in any of the datastores. The locations of
the datastores is obtained from the policies.py file.

This feature can be configured using a property in
archappl.properties. The property is disabled by default.