Fix PlainPBFileNameUtilityTest
Test was failing due to incorrect data logic in the test.
In Monthly test, 12 entries were created but expected November to be the last month. Now creates 11 entries.
In Yearly test, the "curr" variable was tracking the year of the last file + one. So a modification was made to check the year minus one. Similarly with the endTime variable.