Temporarily push out ETL for store based on env variable
Meant for emergencies only.
If you are having trouble with a particular store and it might take some time to fix;
you can restart the archiver with ARCHAPPL_SKIP_ETL_FOR_STORE set to the name of the store.
This will set the initial delay for that store for 10 year into the future.
Note your breathing room for fixing the issue with the storage is really the capacity of the upstream store
and your chances of running out of space with this setting are probably quite high.
But in case of an emergency, you can simply set this environment variable
and etl for this store will be delayed for some time.