Proper cluster startup.
Previously, we would allow you to define any number of appliances in appliances.xml
This would let have spare appliances on standby and then bring them in a hurry if needed.
However; this makes it difficult to detect when the entire cluster ( as opposed to a single appliance ) has started up.
In split brain situations; this can cause havoc especially when you have automated scripts that adds PV's to the archiver.
One of the few ways we can prevent this is to make sure that all appliances listed in appliances.xml have started up and added their PVs to the cluster.
Certain BPL's, most importantly, the archivePV BPL, are blocked till the entire cluster has started up.
Please let us know if this interferes with your deployment and we can come with a different strategy for your installation.