Fixes integration testing by adhoc creating an appliances xml
Recent changes meant that all appliances and no more should be specified in the appliances.xml file. The integration tests were not doing this so the tests were failing. This commit adds an appliances template appliances.xml.j2 which can be used to generate new appliances.xml for arbituary cluster size and configuration. Then adds it to the test configurations.