This is in archappl.properties in addition to the archivePVStartup batch size and sleep seconds
The other change is that we enforce server_ioc_clock_drift on samples with timestamps in the past as well.
This change applies to samples from the second sample onwards.
That is, after we connect to a PV, the first sample is accepted as is.
But from the second sample onwards, if the timestamp is more than clock drift seconds in the past, we drop it
and increment the error counter. The rationale for this is that we'd rather catch IOC's with bad clocks using
a report than by having to look thru ETL logs. Tested this with IOC's using libfaketime and both retrieval and
ETL handled this situation correctly..